+++ src/main/java/net/earlystage/item/BrickBucketItem.java
@@ -5,7 +5,6 @@ import net.minecraft.advancement.criterion.Criteria;
 import net.minecraft.block.Block;
 import net.minecraft.block.BlockState;
 import net.minecraft.block.Blocks;
-import net.minecraft.block.FluidBlock;
 import net.minecraft.block.FluidDrainable;
 import net.minecraft.block.FluidFillable;
 import net.minecraft.entity.Entity;
@@ -17,6 +16,7 @@ import net.minecraft.item.FluidModificationItem;
 import net.minecraft.item.Item;
 import net.minecraft.item.ItemStack;
 import net.minecraft.item.ItemUsage;
+import net.minecraft.item.Items;
 import net.minecraft.particle.ParticleTypes;
 import net.minecraft.registry.tag.FluidTags;
 import net.minecraft.server.network.ServerPlayerEntity;
@@ -37,10 +37,10 @@ import net.minecraft.world.event.GameEvent;
 import org.jetbrains.annotations.Nullable;
 
 @SuppressWarnings("deprecation")
-public class WoodenBucketItem extends Item implements FluidModificationItem {
+public class BrickBucketItem extends Item implements FluidModificationItem {
     private final Fluid fluid;
 
-    public WoodenBucketItem(Fluid fluid, Item.Settings settings) {
+    public BrickBucketItem(Fluid fluid, Item.Settings settings) {
         super(settings);
         this.fluid = fluid;
     }
@@ -48,7 +48,7 @@ public class WoodenBucketItem extends Item implements FluidModificationItem {
     @Override
     public TypedActionResult<ItemStack> use(World world, PlayerEntity user, Hand hand) {
         ItemStack itemStack = user.getStackInHand(hand);
-        BlockHitResult blockHitResult = WoodenBucketItem.raycast(world, user, this.fluid == Fluids.EMPTY ? RaycastContext.FluidHandling.SOURCE_ONLY : RaycastContext.FluidHandling.NONE);
+        BlockHitResult blockHitResult = BrickBucketItem.raycast(world, user, this.fluid == Fluids.EMPTY ? RaycastContext.FluidHandling.SOURCE_ONLY : RaycastContext.FluidHandling.NONE);
         if (blockHitResult.getType() == HitResult.Type.MISS) {
             return TypedActionResult.pass(itemStack);
         }
@@ -61,16 +61,20 @@ public class WoodenBucketItem extends Item implements FluidModificationItem {
             }
             if (this.fluid == Fluids.EMPTY) {
                 BlockState blockState = world.getBlockState(blockPos);
-                if (blockState.getBlock() instanceof FluidDrainable && blockState.getBlock().getFluidState(blockState).isOf(Fluids.WATER) && blockState.get(FluidBlock.LEVEL) == 0) {
-
+                if (blockState.getBlock() instanceof FluidDrainable && !blockState.getFluidState().isEmpty() && blockState.getFluidState().isStill()
+                        && (blockState.getFluidState().isIn(FluidTags.WATER) || blockState.getFluidState().isIn(FluidTags.LAVA))) {
                     user.incrementStat(Stats.USED.getOrCreateStat(this));
-                    user.playSound(SoundEvents.ITEM_BUCKET_FILL, SoundCategory.PLAYERS, 1.0f, 1.0f);
+                    ItemStack itemStack2;
+                    if (blockState.getFluidState().isIn(FluidTags.LAVA)) {
+                        world.playSound(user, blockPos, SoundEvents.ITEM_BUCKET_FILL_LAVA, SoundCategory.BLOCKS, 1.0f, 1.0f);
+                        itemStack2 = new ItemStack(ItemInit.LAVA_BRICK_BUCKET);
+                    } else {
+                        world.playSound(user, blockPos, SoundEvents.ITEM_BUCKET_FILL, SoundCategory.BLOCKS, 1.0f, 1.0f);
+                        itemStack2 = new ItemStack(ItemInit.WATER_BRICK_BUCKET);
+                    }
                     world.emitGameEvent((Entity) user, GameEvent.FLUID_PICKUP, blockPos);
-
-                    ItemStack itemStack2 = new ItemStack(ItemInit.WATER_WOODEN_BUCKET, 1);
                     ItemStack itemStack3 = ItemUsage.exchangeStack(itemStack, user, itemStack2);
-
-                    if (!world.isClient) {
+                    if (!world.isClient()) {
                         world.setBlockState(blockPos, Blocks.AIR.getDefaultState(), Block.NOTIFY_ALL | Block.REDRAW_ON_MAIN_THREAD);
                         Criteria.FILLED_BUCKET.trigger((ServerPlayerEntity) user, itemStack2);
                     }
@@ -86,25 +90,25 @@ public class WoodenBucketItem extends Item implements FluidModificationItem {
                     Criteria.PLACED_BLOCK.trigger((ServerPlayerEntity) user, blockPos3, itemStack);
                 }
                 user.incrementStat(Stats.USED.getOrCreateStat(this));
-                return TypedActionResult.success(WoodenBucketItem.getEmptiedStack(itemStack, user), world.isClient());
+                return TypedActionResult.success(BrickBucketItem.getEmptiedStack(itemStack, user, this.fluid), world.isClient());
             }
             return TypedActionResult.fail(itemStack);
         }
         return TypedActionResult.pass(itemStack);
     }
 
-    public static ItemStack getEmptiedStack(ItemStack stack, PlayerEntity player) {
+    public static ItemStack getEmptiedStack(ItemStack stack, PlayerEntity player, Fluid fluid) {
         if (!player.getAbilities().creativeMode) {
-            return ItemStack.EMPTY;
+            if (fluid == Fluids.LAVA) {
+                return new ItemStack(Items.AIR);
+            }
+            return new ItemStack(ItemInit.BRICK_BUCKET);
         }
         return stack;
     }
 
     @Override
     public void onEmptied(@Nullable PlayerEntity player, World world, ItemStack stack, BlockPos pos) {
-        if (!player.isCreative()) {
-            world.playSound(pos.getX(), pos.getY(), pos.getZ(), SoundEvents.ENTITY_ITEM_BREAK, SoundCategory.PLAYERS, 0.8f, 0.8f + world.random.nextFloat() * 0.4f, false);
-        }
     }
 
     @Override
@@ -134,7 +138,7 @@ public class WoodenBucketItem extends Item implements FluidModificationItem {
             this.playEmptyingSound(player, world, pos);
             return true;
         }
-        if (!world.isClient() && bl && !blockState.isLiquid()) {
+        if (!world.isClient && bl && !blockState.isLiquid()) {
             world.breakBlock(pos, true);
         }
         if (world.setBlockState(pos, this.fluid.getDefaultState().getBlockState(), Block.NOTIFY_ALL | Block.REDRAW_ON_MAIN_THREAD) || blockState.getFluidState().isStill()) {
@@ -147,6 +151,9 @@ public class WoodenBucketItem extends Item implements FluidModificationItem {
     private void playEmptyingSound(@Nullable PlayerEntity player, WorldAccess world, BlockPos pos) {
         SoundEvent soundEvent = this.fluid.isIn(FluidTags.LAVA) ? SoundEvents.ITEM_BUCKET_EMPTY_LAVA : SoundEvents.ITEM_BUCKET_EMPTY;
         world.playSound(player, pos, soundEvent, SoundCategory.BLOCKS, 1.0f, 1.0f);
+        if (this.fluid.isIn(FluidTags.LAVA) && player != null && !player.isCreative()) {
+            world.playSound(player, pos, SoundEvents.ENTITY_GENERIC_BURN, SoundCategory.PLAYERS, 0.5f, 1.0f);
+        }
         world.emitGameEvent((Entity) player, GameEvent.FLUID_PLACE, pos);
     }
 }